requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
python-dotenv==1.0.1
Flask-Migrate==4.0.5 